refator : newProject 다시 만들기 newProject 를 다시 만듭니다. #13
