+++ src/pages/NewProject/newproject.tsx
@@ -4,7 +4,7 @@ import CssBaseline from '@material-ui/core/CssBaseline';
 import Grid from '@material-ui/core/Grid';
 import Container from '@material-ui/core/Container';
 import AppbarShift from '../layout/appbarshift';
-import Stepper from './components/newprojectstepper';
+import NewProjectForm from './components/newProjectForm';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -32,7 +32,7 @@ export default function NewProject(): JSX.Element {
         <div className={classes.appBarSpacer} />
         <Container maxWidth="lg" className={classes.container}>
           <Grid container spacing={2}>
-            <Stepper />
+            <NewProjectForm />
           </Grid>
         </Container>
       </main>
